Improve documentation across different crates
As in title.
