+++ chain/types/src/chain_events.rs
@@ -1,5 +1,6 @@
 //! This module defines types that are related to events that are raised on-chain
 //! in response to actions (transactions, smart contract calls) done by a HOPR node on chain.
+//!
 //! See `chain-actions` and `chain-indexer` crates for details.
 use hopr_crypto_types::types::Hash;
 use hopr_internal_types::prelude::*;
