+++ common/internal-types/src/announcement.rs
@@ -5,7 +5,9 @@ use multiaddr::Multiaddr;
 use std::fmt::{Display, Formatter};
 
 /// Holds the signed binding of the chain key and the packet key.
-/// The signature
+///
+/// The signature is done via the offchain key to bind it with the on-chain key. The structure
+/// then makes it on-chain, making it effectively cross-signed with both keys (offchain and onchain).
 /// This is used to attest on-chain that node owns the corresponding packet key and links it with
 /// the chain key.
 #[derive(Clone, Debug, PartialEq)]
