+++ common/internal-types/src/errors.rs
@@ -2,6 +2,7 @@ use hopr_crypto_types::errors::CryptoError;
 use multiaddr::Error as MultiaddrError;
 use thiserror::Error;
 
+/// Enumeration of all core type related errors.
 #[derive(Error, Debug)]
 pub enum CoreTypesError {
     #[error("{0}")]
