+++ common/internal-types/src/lib.rs
@@ -1,13 +1,20 @@
 //! This crate contains all types that are specific to the HOPR protocol.
 //! As opposed to `hopr-primitive-types` which contains more generic types (not necessarily specific only to HOPR).
 
+/// Contains all types related to node identities.
 pub mod account;
+/// Implements types for ticket acknowledgement.
 pub mod acknowledgement;
+/// Implements types for on-chain announcement of nodes.
 pub mod announcement;
+/// Implements types related to HOPR payment channels.
 pub mod channels;
+/// Enumerates all errors in this crate.
 pub mod errors;
+/// Types related to internal HOPR protocol logic.
 pub mod protocol;
 
+#[doc(hidden)]
 pub mod prelude {
     pub use super::account::*;
     pub use super::acknowledgement::*;
