+++ hopr/hopr-lib/src/lib.rs
@@ -4,7 +4,7 @@
 //! The [`Hopr`] object is standalone, meaning that once it is constructed and run,
 //! it will perform its functionality autonomously. The API it offers serves as a
 //! high level integration point for other applications and utilities, but offers
-//! a complete and fulle featured HOPR node stripped from top level functionality
+//! a complete and fully featured HOPR node stripped from top level functionality
 //! such as the REST API, key management...
 //!
 //! The intended way to use hopr_lib is for a specific tool to be built on top of it,
@@ -13,9 +13,13 @@
 //! For most of the practical use cases, the `hoprd` application should be a preferable
 //! choice.
 
+/// Types specific for blockchain deployments of HOPR.
 mod chain;
+/// Configuration related public types
 pub mod config;
+/// Various public constants.
 pub mod constants;
+/// Enumerates all errors thrown from this library.
 pub mod errors;
 mod helpers;
