+++ hoprd/inbox/src/config.rs
@@ -7,11 +7,9 @@ use validator::{Validate, ValidationError};
 use hopr_internal_types::prelude::*;
 
 pub fn validate_is_power_of_two(value: u32) -> Result<(), ValidationError> {
-    if (value & (value - 1)) != 0 {
-        Err(ValidationError::new("The value is not power of 2"))
-    } else {
-        Ok(())
-    }
+    ((value & (value - 1)) == 0)
+        .then_some(())
+        .ok_or(ValidationError::new("The value is not power of 2"))
 }
 
 /// Holds basic configuration parameters of the `MessageInbox`.
@@ -21,16 +19,22 @@ pub struct MessageInboxConfiguration {
     /// Maximum capacity per-each application tag.
     ///
     /// In the current implementation, the capacity must be a power of two.
+    ///
+    /// Defaults to 512.
     #[validate(custom = "validate_is_power_of_two")]
     #[serde(default = "default_capacity")]
     #[default(default_capacity())]
     pub capacity: u32,
     /// Maximum age of a message held in the inbox until it is purged.
+    ///
+    /// Defaults to 15 minutes.
     #[serde_as(as = "DurationSeconds<u64>")]
     #[serde(default = "just_15_minutes")]
     #[default(just_15_minutes())]
     pub max_age: Duration,
     /// List of tags that are excluded on `push`.
+    ///
+    /// Defaults to \[[DEFAULT_APPLICATION_TAG]\]
     #[serde(default = "default_excluded_tags")]
     #[default(default_excluded_tags())]
     pub excluded_tags: Vec<Tag>,
