+++ logic/path/src/errors.rs
@@ -2,6 +2,7 @@ use hopr_primitive_types::errors::GeneralError;
 use thiserror::Error;
 use utils_db::errors::DbError;
 
+/// Enumerates all errors in this crate.
 #[derive(Error, Debug)]
 pub enum PathError {
     #[error("path is not valid")]
